pyproject.toml
[tool.poetry]
name = "quorial-softwaredev-1"
version = "0.1.0"
description = "QUORIAL - Flask RAG Chat Application"
authors = ["angelinaradovanov <angelina.radovanov@kununu.com>"]
readme = "README.md"
packages = [{include = "src"}, {include = "flask_quorial"}, {include = "run_chat_app.py"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.0"
werkzeug = "^2.3.0"
chromadb = "^1.3.4"
openai = "^2.7.2"
gradio = "^5.33.0"
pandas = "^2.2.3"
numpy = "^1.26.4"
sentence-transformers = "^4.1.0"
tqdm = "^4.67.1"
mistralai = "^0.4.1"
bs4 = "^0.0.2"
langchain-text-splitters = "^1.0.0"
trafilatura = "^2.0.0"
reportlab = "^4.1.0"
pytest = "^9.0.2"
pytest-flask = "^1.3.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
# Main Applications
chat-app = "run_chat_app:main"
gradio-app = "src.app_gradio:main"

# Data Processing Scripts
clean-data = "src.clean_raw_data:main"
preprocess-data = "src.preprocess:main"
chunk-articles = "src.chunking_articles:main"

# ChromaDB Management
rebuild-chroma = "src.rebuild_chroma:main"

# RAG Pipeline
rag-pipeline = "src.rag_pipeline:main"
test-retriever = "src.retriever:main"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
